backend/requirements.txt
fastapi
uvicorn
groq
langgraph
python-dotenv
langgraph-checkpoint-sqlite  
tavily-python